refactor(P2.2): move pattern-comprehension SQL emitters into render_plan/pattern_comprehension_sql.rs
Second Phase-2 module move (REFACTORING_SAFETY_PLAN.md §5.1). The
pattern-comprehension SQL *string*-emitting group — correlated-subquery /
pre-aggregated-CTE renderings for `size([(a)-[:R]->(b) WHERE .. | ..])` /
pattern-count, arrayCount list-comprehension lowering, count(*)- and
arraycount-placeholder substitution passes, the pc edge/node id-column and
CTE-column lookups, plus build_cte_column_map/scan_expr_for_aliases which only
that group uses — is extracted verbatim (31 fns, 2,629 lines) from
plan_builder_utils.rs into a new render_plan/pattern_comprehension_sql.rs.

Pure move, zero logic edits. The only changes are visibility widenings:
12 moved fns + the PcCteResult struct/fields widened to pub(crate); the two
stay-behind helpers quote_qualified_col / extract_from_alias_from_cte_name
widened to pub(crate) so the sibling module can call them. pub(crate) re-exports
from the old path keep the narrow set of existing callers (incl. plan_builder.rs
and the in-file test module) resolving unchanged. Named imports only, no glob,
no import cycle (the new module does not import RenderPlanBuilder).

Per-function byte-diff vs origin/main verified identical modulo those documented
visibility widenings and one benign fmt sig-reflow of emit_array_count_in_subquery.
210 goldens + 1,082-query corpus byte-identical; full cargo test green
(fmt/clippy clean). Ratchet baseline updated for the pure relocation of schema
(from_label_column/to_label_column/type_column) and dialect (Dialect::/databricks)
axis tokens from plan_builder_utils.rs to the new module — all net-zero.

Docs updated in-branch: render_plan/AGENTS.md module map, PRIORITIES.md P-3 +
merge log, REFACTORING_SAFETY_PLAN.md §5.1 row + §9 checklist. D7-rest deferred.
